Check finalizers at reconcile
Finalizers are set when the object should be deleted. Finalizers in a
CRD metadata could be used as a hint for the controller to delete the
dependent objects. Once all the controller actions are completed, the
finalizer should be emptied for k8s to garbage collect it. This fixes
the issue of hang at uninstall by emptying the finalizers in the
controller.