Create dynamic join tokens using node list
With this change, join token is no longer needed to be passed while
creating a new cluster. The operator would query list of nodes and
dynamically create and update join token. This helps in maintaining
up-to-date join tokens which is in sync with the current state of the
cluster. New node IPs are added and deleted node IPs are removed.