Add DaemonSet Job controller
This Job controller creates jobs.batch like tasks but backed by
DaemonSets which can run on all the nodes in the cluster. This can be
used to run certain tasks across all the nodes in the cluster.
Some of the usage of this would be to perform storageos config cleanup,
installing certain dependencies like the kernel modules on all the
nodes, pulling container image on all the nodes, etc.