Create image puller job in upgrade
With this, the new upgrade image is pulled in all the nodes first,
before any upgrade is attempted. Upgrade starts only after the image is
downloaded in all the nodes.

This also changes the jobs spec, adding `HostPath` to set volume path on
host and `MountPath` to set the mount path in the job container.