Add OLM status overview details
This change integrates the resource status details with the OLM status
overview page to show the object status in the web console.

Also adds a new `Members` field in StorageOSCluster status to store the
`Ready` and `Unready` status of the storageos pods. This is used in the
OLM status overview page to show a graph of the cluster status.