Create metadata bundle for openshift scanner
This adds a make target to generate metadata bundle and adds a test for
the generated bundle in the e2e test.
The bundle includes a separate CSV file with container image from rhel
repository and some updated attributes.

Also adds deploy/olm/storageos/storageos.clusterserviceversion.yaml to
maintain the latest version of CSV separately, not modifying the
published CSV.