test/e2e: Add test for CSI helper deployment
Adds a new test TestClusterCSIDeployment to deploy and CSI helper as
Deployment.

Adds operator-sdk-e2e-cleanup() in e2e.sh to cleanup the global
resources created when a cluster is installed. They are not cleaned up
because once the test is over, the operator is deleted immediately by
the test runner, leaving behind the created resources for the cluster.