storageos: Add storageos scheduler deployment
This adds deployment of an external scheduler using kube-scheduler as a
Deployment. This is automatically deployed by default. For a workload to
use storageos scheduler, the pod spec must have `schedulerName` set to
`storageos-scheduler`.

Adds a new StorageOSCluster spec property `disableScheduler` to disable
the scheduler. Also adds `image.hyperkubeContainer` to set the hyperkube
container image to use. By default, gcr.io/google_containers/hyperkube
is the container image is used for hyperkube. The image tag depends on
the version of kubernetes cluster running.