Add missing CR attributes
The StorageOSCluster and NFSServer CRs were not being updated with
default and inferred values. This change adds a check in the reconcile
loop to update the spec if the spec properties need to be updated.

This also fixes the empty StorageClass in the NFSServer CR.

In StorageOSCluster reconcile loop, an spec update requires updating
the current cluster with new properties to reflect any new data in the
cluster deployment.