Allow setting cluster-wide NFS container image
This adds NFSContainer in StorageOSCluster.Images to allow setting the
default NFS container image for a StorageOS cluster. This can be
overridden by specifying a container image in the NFSServer CR.

Also, adds tests for updateSpec() in the NFS controller to test the
inheritance and override of container image and storageclass from
StorageOSCluster into NFS Server.