create etcd secret of the same type as source
This fixes a bug introduced in the refactoring of k8s related code due
to which the etcd secret created for storageos cluster based on a
provided reference secret is of type TLS and not the same type as the
source secret. This results in secret creation failure because TLS
secrets have a strict data field requirements and the source secret
does not have the same data fields.

Adds a test to ensure such issues are caught in the future.