e2e: Add e2e tests for ServiceMonitor
Adds the custom resource scheme for ServiceMonitor to test framework
scheme, to be used with the dynamic client and adds the ServiceMonitor
CRD in deploy/crds/ dir to be included in the e2e test.
The e2e test checks if a ServiceMonitor is created by the operator
when the CRD is discovered in the cluster.