Add leader-for-life leader election
New Operator-SDK projects enable leader-for-life leader election by
default. This can help in running multiple instances of the operator
with only one being the leader until the leader pod is deleted.
This model doesn't affect the admission controller webhook in the
current setup as only the leader pod will be running the webhook and
the other pods will be in a waiting state. When a leader pod is
deleted, the new leader generates new cert for the admission controller
and updates the client cert.