NFS: fix service label selectors
The new k8s labels that were added to all the resources introduced a bug
due to which the Service labels selector for NFS wasn't able to select
any pod, resulting in no endpoints in the service. Due to this NFS
mounts were resulting in connection timeout.
This change fixes the labels by correcting label selectors used in the
Service and introducing a new label app.kubernetes.io/service-for to
define what a service for. The NFS Server service has service-for value
"nfs-server" and the metrics service has the value "nfs-metrics". This
helps easily distinguish between the NFS Server service and the NFS
metrics service. The prometheus service monitor selectors have have been
updated to use the new labels.