webhook handler: use custom decoder
Use a custom recoder in the webhook handler if the decoder is nil. The
decoder may remain nil due to failure in translation of webhook
admission decoder from controller-runtime to the internal copy of the
webhook admission decoder. Since the webhook handler is for pods only
for now, use client-go's scheme to create a new decoder that can decode
pods from a webhook request.
Failure in creating a new decoder will return an internal server error.