Update code to be compatible with new dependencies
There are a lot of breaking changes in controller-runtime v0.2+ and k8s
1.15. This updates all the operator code to adapt to the breaking
changes.

Braking changes include:
- Change in controller-runtime's dynamic client function signature
- Change in controller-runtime's logger name
- Change in controller manager's k8s event recorder function from
mgr.GetRecorder() to mgr.GetEventRecorder().
- admissionregistration/v1beta1 splits Webhook to MutatingWebhook and
ValidatingWebhook. This required updating the copy of
controller-runtime's webhook package to support the new admission API.
- The latest k8s code-generator requires k8s.io/code-generator to be in
the GOPATH. code-generation for clientset can be replaced in the future
using the dynamic client.
- meta/v1 GetOptions no longer has IncludeUninitialized attribute.