Add k8s api resource discovery package
This adds a new package internal/pkg/discovery containing helpers to
check if a resource is supported by the k8s api server. This can be used
in tests to check if the test cluster supports some features before
running the tests. This will also be helpful for CSI deployments in
checking if CSIDriver and CSINode built-in resources are supported in
the cluster and used the built-in resources instead of CSI CRDs for
driver deployment.