Add image env vars in operator deployment manifest
This adds env vars for all the related images used by the operator.
These env vars can be used to set default images used in storageos
deployment. These default images can still be overridden by specifying
images in the StorageOSCluster CR. If no image is specified in the CR or
the operator deployment manifest, a set of fallback images are used
which were the default images before.