Fix local-run make target to run from outside of k8s
local-run used to fail because when the operator is run from outside of
k8s, there's no operator namespace. This changes the fatal error to log
a message and skip creation of service monitor, which requires operator
namespace.
The local-run make target is also updated to disable scheduler because
mutating webhook server runs as part of the operator and will be
inaccessible from within k8s.