Fix unit tests
Update to controller-runtime's fake client broke the unit tests. The
failure errors were:
- "resourceVersion can not be set for Create requests"
- "failed to deploy cluster: secrets '' not found"

The resourceVersion error was occurring because tests with subtests were
creating the same cluster objects. The updated fake client retains the
previously created object. Doing a deepcopy of the object before
creating them in subtests fixed this issue.

The secret not found error was occuring when deploying the legacy driver
where the API secret is updated with a service address. Not sure what
changed in the fake client that's causing this. Since we no longer
support the legacy driver, changing the test to use CSI fixes the tests.